requirements/inference_v2.txt
fastapi==0.85.1
uvicorn==0.19.0
aiohttp==3.8.3
PyYAML==6.0.1
transformers>=4.30.0
torch
bitsandbytes
accelerate
